Move mempool interface impl
