+++ mempool/src/interface/mempool_interface_impl/pool/tests/utils.rs
@@ -19,9 +19,9 @@ use common::primitives::H256;
 use super::*;
 
 #[derive(Debug, PartialEq, Eq, Clone)]
-pub(in crate::mempool_interface_impl::pool::tests) struct ValuedOutPoint {
-    pub(in crate::mempool_interface_impl::pool::tests) outpoint: OutPoint,
-    pub(in crate::mempool_interface_impl::pool::tests) value: Amount,
+pub(in crate::interface::mempool_interface_impl::pool::tests) struct ValuedOutPoint {
+    pub(in crate::interface::mempool_interface_impl::pool::tests) outpoint: OutPoint,
+    pub(in crate::interface::mempool_interface_impl::pool::tests) value: Amount,
 }
 
 impl std::cmp::PartialOrd for ValuedOutPoint {
@@ -53,7 +53,7 @@ fn dummy_output() -> TxOutput {
     TxOutput::new(OutputValue::Coin(value), purpose)
 }
 
-pub(in crate::mempool_interface_impl::pool::tests) fn estimate_tx_size(
+pub(in crate::interface::mempool_interface_impl::pool::tests) fn estimate_tx_size(
     num_inputs: usize,
     num_outputs: usize,
 ) -> usize {
