+++ mempool/src/interface/mod.rs
@@ -14,3 +14,4 @@
 // limitations under the License.
 
 pub mod mempool_interface;
+pub mod mempool_interface_impl;
